Extract dragging of text panel
